Adding programs from Friday's class.
Included are:
Solutions to two versions of exercise 5 of the for loops questions
	- Displaying all even numbers from 10 to 50, with only 5 numbers per line
	- Displaying all even numbers from 50 to 10, with only 5 numbers per line
Example of creating and maintaining a running total (where the value keeps updating as new numbers are added to it)
An open-ended guessing game using a while loop. This will only stop running when the user guesses correctly.